ai-legal-assistant-backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
supabase>=2.3.4
langchain>=0.1.0
langchain-google-genai>=0.0.8
langchain-community>=0.0.12
python-multipart>=0.0.6
PyMuPDF>=1.26.0
python-docx>=1.1.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.5.0
tiktoken>=0.6.0
asyncpg>=0.29.0
pgvector>=0.2.4